Refactoring: Extract rensa trackers from core.
We'd like to organize rensa tracker more, and we'd like to have more
sophisticated tests to rensa tracker.

So, let me move rensa tracker implementation to core/rensa_tracker.
For preventing layer violation, we neeed to have core/rensa_tracker.h
which contain only RensaNonTracker.
core should not depend on core/rensa_tracker.
