+++ src/core/rensa_tracker.h
@@ -1,24 +1,19 @@
 #ifndef CORE_RENSA_TRACKER_H_
 #define CORE_RENSA_TRACKER_H_
 
-#include <glog/logging.h>
-
-#include "base/bmi.h"
 #include "base/unit.h"
 #include "core/field_bits.h"
-#include "core/rensa_result.h"
-#include "core/rensa_track_result.h"
-
-class FieldBits;
 
 // RensaTracker tracks how rensa is vanished.
 // For example, a puyo is vanished in what-th chain, coefficient of each chain, etc.
 // You can pass a RensaTracker to CoreField::simulate() to track the rensa.
-// You can also define you own RensaTracker, and pass it to CoreField::simulate().
+// You can also define your own RensaTracker, and pass it to CoreField::simulate().
 //
-// RensaTracker must define several interface. CoreField::simulate() has several hook poinits
+// RensaTracker must define several interface. CoreField::simulate() has several hook points
 // that calls the corresponding Tracker methods. If you'd like to add a new hook point,
 // you need to define a hook point in CoreField.
+//
+// Here, we define only RensaNonTracker. The other implementations are located on core/rensa_tracker/.
 
 class RensaTrackerBase {
 public:
@@ -43,223 +38,4 @@ public:
 };
 typedef RensaTracker<Unit> RensaNonTracker;
 
-// ----------------------------------------------------------------------
-
-class RensaYPositionTracker : public RensaTrackerBase {
-public:
-    RensaYPositionTracker() :
-        originalY_ {
-            0xFEDCBA9876543210,
-            0xFEDCBA9876543210,
-            0xFEDCBA9876543210,
-            0xFEDCBA9876543210,
-            0xFEDCBA9876543210,
-            0xFEDCBA9876543210,
-            0xFEDCBA9876543210,
-            0xFEDCBA9876543210,
-        }
-    {
-    }
-
-    void track(int /*nthChain*/, int /*numErasedPuyo*/, int /*longBonusCoef*/, int /*colorBonusCoef*/,
-               const FieldBits& vanishedColorPuyoBits, const FieldBits& vanishedOjamaPuyoBits)
-    {
-        const __m128i zero = _mm_setzero_si128();
-        const __m128i ones = _mm_cmpeq_epi8(zero, zero);
-        union {
-            std::uint16_t cols[FieldConstant::MAP_WIDTH];
-            __m128i m;
-        };
-        m = (vanishedColorPuyoBits | vanishedOjamaPuyoBits) ^ ones;
-
-        for (int x = 1; x <= 6; ++x) {
-            originalY_[x] = bmi::extractBits4(originalY_[x], cols[x]);
-        }
-    }
-
-    int originalY(int x, int y) const { return (originalY_[x] >> (4 * y)) & 0xF; }
-
-private:
-    std::uint64_t originalY_[FieldConstant::MAP_WIDTH];
-};
-
-// ----------------------------------------------------------------------
-
-template<>
-class RensaTracker<RensaExistingPositionTrackResult> : public RensaTrackerBase {
-public:
-    RensaTracker(FieldBits bits) : result_(bits) {}
-
-    const RensaExistingPositionTrackResult& result() const { return result_; }
-
-    void trackDrop(FieldBits blender, FieldBits leftOnes, FieldBits rightOnes)
-    {
-        FieldBits m = result_.existingBits();
-
-        __m128i v1 = _mm_and_si128(rightOnes, m);
-        __m128i v2 = _mm_and_si128(leftOnes, m);
-        __m128i v3 = _mm_srli_epi16(v2, 1);
-        __m128i v4 = _mm_or_si128(v1, v3);
-        m = _mm_blendv_epi8(m, v4, blender);
-
-        result_.setExistingBits(m);
-    }
-
-private:
-    RensaExistingPositionTrackResult result_;
-};
-typedef RensaTracker<RensaExistingPositionTrackResult> RensaExistingPositionTracker;
-
-// ----------------------------------------------------------------------
-
-template<>
-class RensaTracker<RensaLastVanishedPositionTrackResult> : public RensaTrackerBase {
-public:
-    const RensaLastVanishedPositionTrackResult& result() const { return result_; }
-
-    void track(int /*nthChain*/, int /*numErasedPuyo*/, int /*longBonusCoef*/, int /*colorBonusCoef*/,
-               const FieldBits& vanishedColorPuyoBits, const FieldBits& vanishedOjamaPuyoBits)
-    {
-        result_.setLastVanishedPositionBits(vanishedColorPuyoBits | vanishedOjamaPuyoBits);
-    }
-
-private:
-    RensaLastVanishedPositionTrackResult result_;
-};
-typedef RensaTracker<RensaLastVanishedPositionTrackResult> RensaLastVanishedPositionTracker;
-
-// ----------------------------------------------------------------------
-
-// RensaTracker<RensaChainTrackResult> tracks in what-th rensa a puyo is vanished.
-template<>
-class RensaTracker<RensaChainTrackResult> : public RensaTrackerBase {
-public:
-    RensaTracker() {}
-
-    const RensaChainTrackResult& result() const { return result_; }
-
-    void track(int nthChain, int numErasedPuyo, int longBonusCoef, int colorBonusCoef,
-               const FieldBits& vanishedColorPuyoBits, const FieldBits& vanishedOjamaPuyoBits)
-    {
-        FieldBits m = (vanishedColorPuyoBits | vanishedOjamaPuyoBits);
-        m.iterateBitPositions([&](int x, int y) {
-            result_.setErasedAt(x, yTracker_.originalY(x, y), nthChain);
-        });
-
-        yTracker_.track(nthChain, numErasedPuyo, longBonusCoef, colorBonusCoef, vanishedColorPuyoBits, vanishedOjamaPuyoBits);
-    }
-
-private:
-    RensaYPositionTracker yTracker_;
-    RensaChainTrackResult result_;
-};
-typedef RensaTracker<RensaChainTrackResult> RensaChainTracker;
-
-// ----------------------------------------------------------------------
-
-template<>
-class RensaTracker<RensaCoefResult> : public RensaTrackerBase {
-public:
-    const RensaCoefResult& result() const { return result_; }
-
-    void track(int nthChain, int numErasedPuyo, int longBonusCoef, int colorBonusCoef,
-               const FieldBits& /*vanishedColorPuyoBits*/, const FieldBits& /*vanishedOjamaPuyoBits*/)
-    {
-        result_.setCoef(nthChain, numErasedPuyo, longBonusCoef, colorBonusCoef);
-    }
-
-private:
-    RensaCoefResult result_;
-};
-typedef RensaTracker<RensaCoefResult> RensaCoefTracker;
-
-// ----------------------------------------------------------------------
-
-template<>
-class RensaTracker<RensaVanishingPositionResult> : public RensaTrackerBase {
-public:
-    RensaTracker() {}
-
-    const RensaVanishingPositionResult& result() const { return result_; }
-
-    void track(int nthChain, int numErasedPuyo, int longBonusCoef, int colorBonusCoef,
-               const FieldBits& vanishedColorPuyoBits, const FieldBits& vanishedOjamaPuyoBits)
-    {
-        vanishedColorPuyoBits.iterateBitPositions([&](int x, int y) {
-            if (yTracker_.originalY(x, y) == y) {
-                result_.setBasePuyo(x, y, nthChain);
-            } else {
-                result_.setFallingPuyo(x, yTracker_.originalY(x, y), y, nthChain);
-            }
-        });
-
-        yTracker_ = RensaYPositionTracker();
-        yTracker_.track(nthChain, numErasedPuyo, longBonusCoef, colorBonusCoef, vanishedColorPuyoBits, vanishedOjamaPuyoBits);
-    }
-
-private:
-    RensaVanishingPositionResult result_;
-    RensaYPositionTracker yTracker_;
-};
-typedef RensaTracker<RensaVanishingPositionResult> RensaVanishingPositionTracker;
-
-// ----------------------------------------------------------------------
-
-template<typename Tracker1, typename Tracker2>
-class CompositeRensaTracker : public RensaTrackerBase {
-public:
-    CompositeRensaTracker(Tracker1* tracker1, Tracker2* tracker2) :
-        tracker1_(tracker1), tracker2_(tracker2)
-    {
-    }
-
-    void track(int nthChain, int numErasedPuyo, int longBonusCoef, int colorBonusCoef,
-               const FieldBits& vanishedColorPuyoBits, const FieldBits& vanishedOjamaPuyoBits)
-    {
-        tracker1_->track(nthChain, numErasedPuyo, longBonusCoef, colorBonusCoef,
-                         vanishedColorPuyoBits, vanishedOjamaPuyoBits);
-        tracker2_->track(nthChain, numErasedPuyo, longBonusCoef, colorBonusCoef,
-                         vanishedColorPuyoBits, vanishedOjamaPuyoBits);
-    }
-
-    void trackDrop(FieldBits blender, FieldBits leftOnes, FieldBits rightOnes)
-    {
-        tracker1_->trackDrop(blender, leftOnes, rightOnes);
-        tracker2_->trackDrop(blender, leftOnes, rightOnes);
-    }
-
-private:
-    Tracker1* tracker1_;
-    Tracker2* tracker2_;
-};
-
-// ----------------------------------------------------------------------
-
-// This is the same as RensaChainTracker, however, the result is passed as pointer.
-class RensaChainPointerTracker : public RensaTrackerBase {
-public:
-    explicit RensaChainPointerTracker(RensaChainTrackResult* trackResult) :
-        result_(trackResult)
-    {
-        // TODO(mayah): Assert trackResult is initialized?
-    }
-
-    const RensaChainTrackResult& result() const { return *result_; }
-
-    void track(int nthChain, int numErasedPuyo, int longBonusCoef, int colorBonusCoef,
-               const FieldBits& vanishedColorPuyoBits, const FieldBits& vanishedOjamaPuyoBits)
-    {
-        FieldBits m = (vanishedColorPuyoBits | vanishedOjamaPuyoBits);
-        m.iterateBitPositions([&](int x, int y) {
-            result_->setErasedAt(x, yTracker_.originalY(x, y), nthChain);
-        });
-
-        yTracker_.track(nthChain, numErasedPuyo, longBonusCoef, colorBonusCoef, vanishedColorPuyoBits, vanishedOjamaPuyoBits);
-    }
-
-private:
-    RensaYPositionTracker yTracker_;
-    RensaChainTrackResult* result_;
-};
-
-#endif
+#endif // CORE_RENSA_TRACKER_H_
