+++ src/core/rensa_tracker/rensa_track_result_test.cc
@@ -1,4 +1,4 @@
-#include "core/rensa_track_result.h"
+#include "core/rensa_tracker/rensa_track_result.h"
 
 #include <gtest/gtest.h>
