+++ src/cpu/CMakeLists.txt
@@ -15,6 +15,7 @@ function(cpu_target_link_common_libraries target)
     target_link_libraries(${target}_${CPU_NAMESPACE} puyoai_core_httpd)
     target_link_libraries(${target}_${CPU_NAMESPACE} ${LIB_MICROHTTPD})
   endif()
+  target_link_libraries(${target}_${CPU_NAMESPACE} puyoai_core_rensa_tracker)
   target_link_libraries(${target}_${CPU_NAMESPACE} puyoai_core)
   target_link_libraries(${target}_${CPU_NAMESPACE} puyoai_base)
   target_link_libraries(${target}_${CPU_NAMESPACE} ${LIB_JSONCPP})
