+++ src/duel/CMakeLists.txt
@@ -15,6 +15,7 @@ endif()
 target_link_libraries(duel puyoai_core_server)
 target_link_libraries(duel puyoai_core_algorithm)
 target_link_libraries(duel puyoai_core_server_connector)
+target_link_libraries(duel puyoai_core_rensa_tracker)
 target_link_libraries(duel puyoai_core)
 target_link_libraries(duel puyoai_base)
 target_link_libraries(duel ${LIB_JSONCPP})
