+++ src/wii/CMakeLists.txt
@@ -17,6 +17,7 @@ function(puyoai_wii_add_executable exe)
     target_link_libraries(${exe} puyoai_core_server_connector)
     target_link_libraries(${exe} puyoai_core_server)
     target_link_libraries(${exe} puyoai_core_algorithm)
+    target_link_libraries(${exe} puyoai_core_rensa_tracker)
     target_link_libraries(${exe} puyoai_core)
     target_link_libraries(${exe} puyoai_gui)
     target_link_libraries(${exe} puyoai_base)
